# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Warehouse Designer API",
    "description": "Authoritative layout compiler and persistence for the 3D warehouse designer. The TypeScript compiler in packages/layout-core is the interactive mirror of the compiler used here; both are pinned by fixtures/layout-conformance.",
    "version": "0.0.0"
  },
  "paths": {
...
    "/api/warehouses": {
      "get": {
        "tags": [
          "warehouses"
        ],
        "summary": "List Warehouses",
        "operationId": "list_warehouses_api_warehouses_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WarehouseOut"
                  },
                  "type": "array",
                  "title": "Response List Warehouses Api Warehouses Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "warehouses"
        ],
        "summary": "Create Warehouse",
        "description": "Create a warehouse, optionally seeding it with a document.\n\nWhen ``doc`` is supplied it wins outright \u2014 including its own footprint \u2014 because\nthe document is the source of truth and a payload whose footprint disagreed with\nits document would otherwise create two conflicting records.",
        "operationId": "create_warehouse_api_warehouses_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WarehouseCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WarehouseOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/warehouses/{warehouse_id}": {
      "get": {
        "tags": [
          "warehouses"
        ],
        "summary": "Get Warehouse Detail",
        "operationId": "get_warehouse_detail_api_warehouses__warehouse_id__get",
        "parameters": [
          {
            "name": "warehouse_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Warehouse Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WarehouseOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/warehouses/{warehouse_id}/draft": {
      "put": {
        "tags": [
          "warehouses"
        ],
        "summary": "Put Draft",
        "description": "Autosave. Debounced by the client; guarded by the revision token here.",
        "operationId": "put_draft_api_warehouses__warehouse_id__draft_put",
        "parameters": [
          {
            "name": "warehouse_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Warehouse Id"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DraftPut"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "warehouses"
        ],
        "summary": "Discard Draft",
        "description": "Throw the draft away, keeping any published version intact.\n\nA discard bumps the revision too: a client that still holds the old token must\nfind out that its working copy is no longer the server's.",
        "operationId": "discard_draft_api_warehouses__warehouse_id__draft_delete",
        "parameters": [
          {
            "name": "warehouse_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Warehouse Id"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/warehouses/{warehouse_id}/validate": {
      "post": {
        "tags": [
          "warehouses"
        ],
        "summary": "Validate Warehouse Layout",
        "description": "Compile without writing. The editor's pre-flight for a save or a publish.",
        "operationId": "validate_warehouse_layout_api_warehouses__warehouse_id__validate_post",
        "parameters": [
          {
            "name": "warehouse_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Warehouse Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Validate Warehouse Layout Api Warehouses  Warehouse Id  Validate Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/warehouses/{warehouse_id}/publish": {
      "post": {
        "tags": [
          "warehouses"
        ],
        "summary": "Publish Warehouse Layout",
        "description": "Validate, materialise and publish in one transaction.\n\nIdempotent on ``doc_hash``: publishing the document that is already live returns\nthe existing version with ``changed: false``.",
        "operationId": "publish_warehouse_layout_api_warehouses__warehouse_id__publish_post",
        "parameters": [
          {
            "name": "warehouse_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Warehouse Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/warehouses/{warehouse_id}/versions": {
      "get": {
        "tags": [
          "warehouses"
        ],
        "summary": "List Versions",
        "operationId": "list_versions_api_warehouses__warehouse_id__versions_get",
        "parameters": [
          {
            "name": "warehouse_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Warehouse Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VersionOut"
                  },
                  "title": "Response List Versions Api Warehouses  Warehouse Id  Versions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/warehouses/{warehouse_id}/layout": {
      "get": {
        "tags": [
          "warehouses"
        ],
        "summary": "Get Published Layout",
        "description": "The published layout for Operate mode.\n\nBins are read from the materialised rows rather than recompiled, because those\nrows carry the surrogate ids that placements reference. The document is also\nrecompiled and its hash returned alongside, so a snapshot that no longer matches\nwhat the compiler would produce today is visible as ``conflicts`` instead of\nbeing silently served.",
        "operationId": "get_published_layout_api_warehouses__warehouse_id__layout_get",
        "parameters": [
          {
            "name": "warehouse_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Warehouse Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishedLayoutOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DraftPut": {
        "properties": {
          "doc": {
            "additionalProperties": true,
            "type": "object",
            "title": "Doc"
          },
          "clientDocHash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Clientdochash"
          }
        },
        "type": "object",
        "required": [
          "doc"
        ],
        "title": "DraftPut"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "WarehouseCreate": {
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "default": ""
          },
          "lengthM": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Lengthm"
          },
          "widthM": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Widthm"
          },
          "heightM": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Heightm"
          },
          "origin": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "number"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Origin"
          },
          "doc": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Doc"
          }
        },
        "type": "object",
        "required": [
          "code",
          "lengthM",
          "widthM",
          "heightM"
        ],
        "title": "WarehouseCreate",
        "description": "A warehouse starts from a footprint, with an optional first document.\n\nSupplying ``doc`` lets the client seed a new warehouse from the layout it already\nhas open instead of round-tripping a blank one through a draft save."
      },
      "WarehouseOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "lengthM": {
            "type": "number",
            "title": "Lengthm"
          },
          "widthM": {
            "type": "number",
            "title": "Widthm"
          },
          "heightM": {
            "type": "number",
            "title": "Heightm"
          },
          "draftRevision": {
            "type": "integer",
            "title": "Draftrevision"
          },
          "draft": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VersionOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "published": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VersionOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "code",
          "name",
          "lengthM",
          "widthM",
          "heightM",
          "draftRevision"
        ],
        "title": "WarehouseOut"
      }
    }
  }
}